requirement.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2
Django-Verify-Email==0.0.4
importlib-metadata==3.4.0
kombu==5.0.2
prompt-toolkit==3.0.10
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
